Tidy up modal/dropdown: remove errorMessage
